+++ example/lib/native_app.dart
@@ -10,7 +10,7 @@ import 'pages/button_page.dart';
 import 'pages/checkbox_page.dart';
 import 'pages/chip_page.dart';
 import 'pages/color_page.dart';
-import 'pages/group_page.dart';
+import 'pages/container_page.dart';
 import 'pages/header_box_page.dart';
 import 'pages/header_page.dart';
 import 'pages/icon_page.dart';
@@ -83,7 +83,7 @@ class MyApp extends StatelessWidget {
                         child: ThemeModeSegmentedButton(),
                       ),
                       const SBBListHeader('Basics'),
-                      SBBGroup(
+                      SBBContentBox(
                         child: Column(
                           mainAxisAlignment: MainAxisAlignment.center,
                           children: <Widget>[
@@ -95,7 +95,7 @@ class MyApp extends StatelessWidget {
                       ),
                       const SizedBox(height: sbbDefaultSpacing),
                       const SBBListHeader('Elements'),
-                      SBBGroup(
+                      SBBContentBox(
                         child: Column(
                           mainAxisAlignment: MainAxisAlignment.center,
                           children: const <Widget>[
@@ -123,12 +123,12 @@ class MyApp extends StatelessWidget {
                       ),
                       const SizedBox(height: sbbDefaultSpacing),
                       const SBBListHeader('Modules'),
-                      SBBGroup(
+                      SBBContentBox(
                         child: Column(
                           mainAxisAlignment: MainAxisAlignment.center,
                           children: const <Widget>[
                             _DemoEntry('Autocompletion', AutocompletionPage()),
-                            _DemoEntry('Group', GroupPage()),
+                            _DemoEntry('Group', ContainerPage()),
                             _DemoEntry('Header', HeaderPage()),
                             _DemoEntry('Headerbox', HeaderBoxPage()),
                             _DemoEntry('Modal', ModalPage()),