+++ example/lib/pages/checkbox_page.dart
@@ -79,7 +79,7 @@ class CheckboxPageState extends State<CheckboxPage> {
               ),
               const SizedBox(height: sbbDefaultSpacing),
               const SBBListHeader('Checkbox Item - List'),
-              SBBGroup(
+              SBBContentBox(
                 child: Column(
                   children: [
                     SBBCheckboxListItem(
@@ -147,7 +147,7 @@ class CheckboxPageState extends State<CheckboxPage> {
               Column(
                 spacing: sbbDefaultSpacing * 0.5,
                 children: [
-                  SBBGroup(
+                  SBBContentBox(
                     child: SBBCheckboxListItem.boxed(
                       value: _listItemValue1,
                       label: 'Label',
@@ -155,7 +155,7 @@ class CheckboxPageState extends State<CheckboxPage> {
                       onChanged: _isEnabled ? (value) => setState(() => _listItemValue1 = value) : null,
                     ),
                   ),
-                  SBBGroup(
+                  SBBContentBox(
                     child: SBBCheckboxListItem.boxed(
                       value: _listItemValue2,
                       label: 'Tristate',
@@ -163,7 +163,7 @@ class CheckboxPageState extends State<CheckboxPage> {
                       onChanged: _isEnabled ? (value) => setState(() => _listItemValue2 = value) : null,
                     ),
                   ),
-                  SBBGroup(
+                  SBBContentBox(
                     child: SBBCheckboxListItem.boxed(
                       value: _listItemValue4,
                       label: 'Leading Icon',
@@ -171,7 +171,7 @@ class CheckboxPageState extends State<CheckboxPage> {
                       leadingIcon: SBBIcons.alarm_clock_small,
                     ),
                   ),
-                  SBBGroup(
+                  SBBContentBox(
                     child: SBBCheckboxListItem.boxed(
                       value: _listItemValue5,
                       label: 'Leading and Trailing Icon',
@@ -180,7 +180,7 @@ class CheckboxPageState extends State<CheckboxPage> {
                       trailingIcon: SBBIcons.dog_small,
                     ),
                   ),
-                  SBBGroup(
+                  SBBContentBox(
                     child: SBBCheckboxListItem.boxed(
                       value: _listItemValue3,
                       label: 'Button',
@@ -189,7 +189,7 @@ class CheckboxPageState extends State<CheckboxPage> {
                       onCallToAction: () => sbbToast.show(title: 'Button pressed'),
                     ),
                   ),
-                  SBBGroup(
+                  SBBContentBox(
                     child: SBBCheckboxListItem.custom(
                       value: _listItemValue6,
                       label: 'Custom trailing Widget',
@@ -200,7 +200,7 @@ class CheckboxPageState extends State<CheckboxPage> {
                       ),
                     ),
                   ),
-                  SBBGroup(
+                  SBBContentBox(
                     child: SBBCheckboxListItem.boxed(
                       value: _listItemValue7,
                       label: 'Multiline Label with\nSecondary Label',
@@ -211,7 +211,7 @@ class CheckboxPageState extends State<CheckboxPage> {
                       onChanged: _isEnabled ? (value) => setState(() => _listItemValue7 = value) : null,
                     ),
                   ),
-                  SBBGroup(
+                  SBBContentBox(
                     child: SBBCheckboxListItem.boxed(
                       value: _listItemValue8,
                       label: 'Loading',