+++ example/lib/pages/container_page.dart
@@ -3,8 +3,8 @@ import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 
 import '../native_app.dart';
 
-class GroupPage extends StatelessWidget {
-  const GroupPage({super.key});
+class ContainerPage extends StatelessWidget {
+  const ContainerPage({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -18,16 +18,16 @@ class GroupPage extends StatelessWidget {
       children: <Widget>[
         ThemeModeSegmentedButton(),
         SizedBox(height: sbbDefaultSpacing * 2),
-        SBBGroup(child: content('Default')),
+        SBBContentBox(child: content('Default')),
         SizedBox(height: sbbDefaultSpacing),
-        SBBGroup(color: SBBColors.royal, child: content('Different Color')),
+        SBBContentBox(color: SBBColors.royal, child: content('Different Color')),
         SizedBox(height: sbbDefaultSpacing),
-        SBBGroup(
+        SBBContentBox(
           padding: EdgeInsets.symmetric(vertical: sbbDefaultSpacing),
           child: content('Extra padding'),
         ),
         SizedBox(height: sbbDefaultSpacing),
-        SBBGroup(margin: EdgeInsets.all(sbbDefaultSpacing * 4), child: content('Extra margin')),
+        SBBContentBox(margin: EdgeInsets.all(sbbDefaultSpacing * 4), child: content('Extra margin')),
       ],
     );
   }