+++ lib/src/container/container.dart
@@ -1,0 +1,3 @@
+export 'sbb_content_box.dart';
+export 'theme/sbb_content_box_style.dart';
+export 'theme/sbb_content_box_theme_data.dart';