+++ lib/src/container/sbb_content_box.dart
@@ -1,17 +1,20 @@
 import 'package:flutter/material.dart';
+import 'package:sbb_design_system_mobile/src/container/container.dart';
 
 import '../../sbb_design_system_mobile.dart';
 
-/// SBBGroup. Use to structure content.
+@Deprecated('Use SBBContentBox.')
+typedef SBBGroup = SBBContentBox;
+
+/// SBBContentBox. Use to structure content.
 ///
 /// May be placed anywhere in the screen to structure other components.
 ///
-/// Referred to as `Container` component in the [SBB Design System](https://digital.sbb.ch/de/design-system/mobile/components/container/)
-///
 /// See also:
 /// - [SBBHeaderbox]: A variant of this component fixed below the [SBBHeader].
-class SBBGroup extends StatelessWidget {
-  const SBBGroup({
+/// - [Figma Design Specification](https://www.figma.com/design/ZBotr4yqcEKqqVEJTQfSUa/Design-System-Mobile?node-id=141-2024)
+class SBBContentBox extends StatelessWidget {
+  const SBBContentBox({
     super.key,
     this.margin,
     this.padding,
@@ -21,21 +24,21 @@ class SBBGroup extends StatelessWidget {
     required this.child,
   });
 
-  /// The empty space that surrounds the [SBBContentbox].
+  /// The empty space that surrounds the [SBBContentBox].
   ///
-  /// If this property is null then [SBBGroupStyle.margin] is used, which defaults to
+  /// If this property is null then [SBBContentBoxStyle.margin] is used, which defaults to
   /// [EdgeInsets.zero].
   final EdgeInsetsGeometry? margin;
 
-  /// The empty space that separates the [child] and the edge of [SBBContentbox].
+  /// The empty space that separates the [child] and the edge of [SBBContentBox].
   ///
-  /// If this property is null then [SBBGroupStyle.padding] is used, which defaults to
+  /// If this property is null then [SBBContentBoxStyle.padding] is used, which defaults to
   /// [EdgeInsets.zero].
   final EdgeInsetsGeometry? padding;
 
   /// The box's background color.
   ///
-  /// If this property is null then the ambient [SBBGroupStyle.color] is used, which is
+  /// If this property is null then the ambient [SBBContentBoxStyle.color] is used, which is
   /// by default:
   /// * [SBBColors.white] in light mode
   /// * [SBBColors.charcoal] in dark mode
@@ -43,14 +46,14 @@ class SBBGroup extends StatelessWidget {
 
   /// {@macro flutter.material.Material.clipBehavior}
   ///
-  /// If this property is null then [SBBGroupStyle.clipBehavior] is used, which is by
+  /// If this property is null then [SBBContentBoxStyle.clipBehavior] is used, which is by
   /// default [Clip.hardEdge].
   final Clip? clipBehavior;
 
   /// Whether this widget represents a single semantic container, or if false
   /// a collection of individual semantic nodes.
   ///
-  /// If null, will use the value of [SBBGroupStyle.isSemanticContainer], which is by default
+  /// If null, will use the value of [SBBContentBoxStyle.isSemanticContainer], which is by default
   /// true.
   ///
   /// Setting this flag to true will attempt to merge all child semantics into
@@ -66,7 +69,11 @@ class SBBGroup extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final SBBGroupStyle style = SBBGroupStyle.of(context);
+    // TODO (@smallTrogdor): remove this with version 5.0.0
+    final SBBContentBoxStyle legacyStyle = SBBContentBoxStyle.of(context);
+    final SBBContentBoxStyle newStyle = Theme.of(context).sbbContentBoxTheme!.style!;
+
+    final style = legacyStyle.merge(newStyle);
 
     return Semantics(
       container: isSemanticContainer ?? style.isSemanticContainer!,