+++ lib/src/picker/sbb_date_time_picker.dart
@@ -100,7 +100,7 @@ class SBBDateTimePicker extends StatefulWidget {
         children: [
           Padding(
             padding: const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing),
-            child: SBBGroup(
+            child: SBBContentBox(
               child: SBBDateTimePicker(
                 initialDateTime: initialDateTime,
                 minimumDateTime: minimumDateTime,