+++ lib/src/theme/styles/styles.dart
@@ -1,10 +1,10 @@
+export '../../container/theme/sbb_content_box_style.dart';
 export '../../button/theme/sbb_button_style.dart';
 export 'src/promotion_box_style.dart';
 export 'src/sbb_base_style.dart';
 export 'src/sbb_basic_control_style.dart';
 export 'src/sbb_control_style.dart';
 export 'src/sbb_control_styles.dart';
-export 'src/sbb_group_style.dart';
 export 'src/sbb_headerbox_style.dart';
 export 'src/sbb_list_item_style.dart';
 export 'src/sbb_pagination_style.dart';