+++ test/sbb_modal_popup_test.dart
@@ -30,7 +30,7 @@ class ModalPopupTest extends StatelessWidget {
     mainAxisSize: MainAxisSize.min,
     children: [
       const SBBListHeader('Modal Popup'),
-      SBBGroup(
+      SBBContentBox(
         child: Padding(
           padding: const EdgeInsets.all(sbbDefaultSpacing * .5),
           child: Column(