+++ CHANGELOG.md
@@ -38,6 +38,7 @@ It is expected that you keep this format strictly, since we depend on it in our 
 - BREAKING: `SBBChip`: renamed `badgeText` to `trailingText` - see [v5 migration guide]
 - BREAKING: `SBBRadio` needs an `SBBRadioGroup` ancestor for its group value and `onChanged` callback - see [v5 migration guide]
 - dropped support for Flutter SDK 3.32.x
+- BREAKING: removed default left margin on component `SBBTextField` (#402)
 
 ### Removed
 
@@ -51,6 +52,22 @@ It is expected that you keep this format strictly, since we depend on it in our 
 - Removed `text` of `SBBStatus` - see [v5 migration guide]
 - Removed `onChanged` & `groupValue` from `SBBRadio` & `SBBRadioListItem` - see [v5 migration guide]
 - Removed `padding` from `SBBRadio` - moved to `SBBRadioStyle` - see [v5 migration guide]
+
+
+## [4.4.0] - 2025-12-18
+
+### Added
+
+- Added `SBBContentBox` and `SBBContentBoxThemeData` with `SBBContentBoxStyle` to align with design specifications
+
+### Changed
+
+- (auto): updated icon lib to version 1.8.14
+
+### Deprecated
+
+- `SBBGroup`: use `SBBContentBox` instead
+- `SBBGroupStyle`: use `SBBContentBoxStyle` instead
 
 ## [4.3.1] - 2025-11-21
 
@@ -77,7 +94,6 @@ It is expected that you keep this format strictly, since we depend on it in our 
 ### Changed
 
 - (auto): updated icon lib to version 1.8.11
-- BREAKING: removed default left margin on component `SBBTextField` (#402)
 
 ### Deprecated
 
@@ -723,7 +739,9 @@ It is expected that you keep this format strictly, since we depend on it in our 
 [v5 migration guide]: migration_guide.md
 
 
-[Unreleased]: https://github.com/SchweizerischeBundesbahnen/design_system_flutter/compare/4.3.1...HEAD
+[Unreleased]: https://github.com/SchweizerischeBundesbahnen/design_system_flutter/compare/4.4.0...HEAD
+
+[4.4.0]: https://github.com/SchweizerischeBundesbahnen/design_system_flutter/compare/4.3.1...4.4.0
 
 [4.3.1]: https://github.com/SchweizerischeBundesbahnen/design_system_flutter/compare/4.3.0...4.3.1